pyproject.toml
# SPDX-License-Identifier: MIT

[tool.poetry]
name = "avnie"
version = "2024.7.28"
description = "A fast & user-friendly command-line interface (CLI) for avro.py"
authors = ["HitBlast <hitblastlive@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
avnie = "avnie.cli:run"

[tool.poetry.dependencies]
python = "^3.8"
avro-py = "^2024.7.13"
typer = "^0.12.3"
requests = "^2.32.3"
pyclip = "^0.7.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"
pytest = "^8.3.2"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
